refactor: renderer structure
